[REFACTOR] 파일 줄이기 #109
