--- Assets/Scripts/SLAM/Detection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a45b6a0266311df44ad9df1e3967f8b4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
