--- Assets/Scripts/SLAM/SlamQuiizUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f20f39146e3de6047a3b2c62990f02fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
